no more data addition
any additional data that needs to be added can be configured as a json
file, hence no hardcoding / populating the db on the backend
