+++ text/pasto.json
@@ -1,0 +1,20 @@
+{
+  "SecurityNote": "These terms are for demonstration purpose only. The current project does not correspond to an actual issued security and offered by a registered broker dealer. If it was, this note would provide legal disclaimers and notes about the security asset.",
+  "Purpose": "Proceeds from this project's raise are invested in the development of a Proof-of-Concept installation in the Segunda Unidad Pasto School in Aibonito, Puerto Rico. The PoC is part of research initiative of the Yale OpenLab and the Digital Currency Initiative of the MIT Media Lab.",
+  "Description": "The public school of SU Pasto is a Montessori-based school located in the town of Aibonito, Puerto Rico and operated by the Instituto Nueva Escuela. After hurricanes Maria and Irma affected the island of Puerto Rico, the school was out of power for over 6 months. However, it operated as a defacto community shelter since its Parent-Teacher Association is a strong cohesive group in the Aibonito community. The school had an emergency generator and was able to cater to community, needs during the aftermath. The most important aspect missing for the school was a resilient telecommunication system to connect with first responders. The current project entails the installation of a 1kW system connected to the administrative building’s subpanel. The solar system powers critical loads in the admin building, including emergency lights, a telecommunication system, the principal's computer and a mini fridge for first aids. The project is a proof-of-concept installation design and developed by the Yale Openlab and the DCI at the MIT Media Lab. The object is to test the interconnection of the system and the internet connected sensors powering simulated payment mechanisms. ",
+  "Context": "One and a half years after hurricane Maria hit the island, schools are still exposed to a centralized and high-carbon energy system vulnerable to climate impacts. At the same time, most schools are becoming grass roots community centers facilitating discussions of concerned parents on how to increase climate & social resilience in the whole community. The Puerto Rican (PR) government and the department of education are working to appoint schools as emergency shelters —nodes with, robust energy and communication systems— for the community to reach out in the event of unavoidable climate shocks. Financing is a key gap.This project acts as a proof-of-concept finance mechanism and school that be one of emergency shelter school appointed by the government. Its strong role in the Aibonito community also makes it eligible for the solar system to expand into a community microgrid system",
+  "Design Description": "The installation was performed over the administration building where the principal office and the representative of the Department of Education sit.. The school has 14 buildings connected to a 300kW substation. The solar system is expected to cover 20% of the admin building for critical loads only.",
+  "Installation Archetype": "This is a model installation in that the solar and battery support only a subpanel of the building circuitry, where only critical loads have been connected. The system is configured as a grid-tied installation, in that the main grid can also support other loads in the panel, and can also be used to power the battery bank. The installation allows for the interconnection of an emergency generator but it is currently not permanently linked to the system. The system, size can be expanded from 1 kW up to 5kW to cover a greater % of the building load.",
+  "IT Infrastructure": "Different layers of IoT energy monitoring devices informs of the energy generated by the solar system and the energy consumed by the end-user. This energy data that will eventually use an intermediary oracle, activates payment transactions per kWh at a specified tariff rate and at specific intervals. Energy payment transactions occur between the end-users wallet/account and a main escrow smart contract that accrues the funds and issues payment of interests (i.e. coupons), to investors. Once the accrued funds can cover the original principle, the contracts pays off investors and the bonds matures.",
+  "Consultation": "The MIT and Yale team convened a meeting with the Parent-Teacher Organisation, led by principal Janice Alejandro. Over 50 members of the community gathered to discuss the project, with unanimous approval and significant interest for its replication. ",
+  "Participation": "The principal, deputy principal, janitor, two teachers and one parent have volunteered as caretakers of the system to monitor its status, report any qualitative information and coordinate with the operation & maintenance required. ",
+  "Outreach": "The system has been installed with instructions so that it can act as a pedagogical site for students to learn about the merits of solar energy, electricity and basic electronics. Talk about solar energy will be given every semester in the context of climate change communication to the community. ",
+  "Governance": "The board of the Parent-Teacher Organisation, composed of 5 chairs, will coordinate with the director of the Yale Openlab and the head of strategy at the MIT DCI to define any governance steps that the system requires. ",
+  "BUSINESS NUMBERS AND PAYMENT SCHEME": "The bond issued by the PR Department of Education covered the principle cost, used for labor and materials. The PPA revenue accrues to pay coupons and mature the bond. The MIT is registered as a first-loss guarantor. ",
+  "Capital Expenditure": "The expected capital cost of the project were met with a superavit of $3000 thanks to the low-cost labor provided by Genmoji Solar. ",
+  "Project Revenue": "The Department of Education of the Puerto Rican government is legally responsible to cover 100% of school at an agreed rate of 0.24 $/kWh. This as a PPA consists of the project revenue.",
+  "Project Expenses": "The project has an allocated O&M (Operation & Management) expense of $1000 every six months with roll-overs. ",
+  "Net Income": "At the present rate of payments, the DofEd is expected to pay off the bond and become full owner of the system by 2025.",
+  "Documents and Contracts": "Below are the links to the contracts associated to the project. Paper based legal contract, agreements and presentation are stored in a decentralised database, and their digital fingerprint (or hash) are stored in the blockchain ledger of the project to ensure immutability. Click on the icons below to access the documents. Some contract documents are self executed by blockchain-based smart contracts that you can access below. ",
+  "HOW TO REVIEW BLOCKCHAIN BASED SMART CONTRACTS": "The current smart contracts are written in the Go scripting language. Function names relate to specific terms or schemes of the contract, and often structured as ‘if, then’ statements. Contract descriptions in plain english are provided in grey letter prior to each function.  To perform a careful revision of smart contracts, you need a certified smart contract auditor. This project’s have been audited and approved by all parties"
+}
